공장 설비 클릭 이벤트 추가 - 미니맵 설비 하이라이드 - 카메라 lookAt 이동
